rename InputRequirements
