+++ tests/integration/pipelex/pipes/controller/pipe_parallel/test_pipe_parallel_validation.py
@@ -1,7 +1,7 @@
 from pipelex.core.concepts.concept_blueprint import ConceptBlueprint
 from pipelex.core.concepts.concept_factory import ConceptFactory
-from pipelex.core.pipes.pipe_input import PipeInput
-from pipelex.core.pipes.pipe_input_blueprint import InputRequirementBlueprint
+from pipelex.core.pipes.input_requirement_blueprint import InputRequirementBlueprint
+from pipelex.core.pipes.input_requirements import InputRequirements
 from pipelex.hub import get_concept_provider, get_pipe_provider
 from pipelex.pipe_controllers.parallel.pipe_parallel_blueprint import PipeParallelBlueprint
 from pipelex.pipe_controllers.parallel.pipe_parallel_factory import PipeParallelFactory
@@ -195,7 +195,7 @@ class TestPipeParallelValidation:
         needed_inputs = pipe_parallel.needed_inputs()
 
         # Verify it returns a PipeInput object
-        assert isinstance(needed_inputs, PipeInput)
+        assert isinstance(needed_inputs, InputRequirements)
         assert hasattr(needed_inputs, "root")
         assert isinstance(needed_inputs.root, dict)
         # With no sub-pipes, should return empty inputs
