+++ tests/unit/pipelex/core/memory/test_dry_working_memory.py
@@ -4,7 +4,7 @@ from pipelex import log
 from pipelex.core.concepts.concept_factory import ConceptFactory
 from pipelex.core.concepts.concept_native import NativeConceptEnum
 from pipelex.core.memory.working_memory_factory import WorkingMemoryFactory
-from pipelex.core.pipes.pipe_input import TypedNamedInputRequirement
+from pipelex.core.pipes.input_requirements import TypedNamedInputRequirement
 from pipelex.core.stuffs.stuff_content import PageContent, TextContent
 from tests.test_pipelines.tricky_questions import ThoughtfulAnswer
