+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/pipe_operator/pipe_func/test_data.py
@@ -1,6 +1,6 @@
 from typing import ClassVar
 
-from pipelex.core.pipes.pipe_input_blueprint import InputRequirementBlueprint
+from pipelex.core.pipes.input_requirement_blueprint import InputRequirementBlueprint
 from pipelex.libraries.pipelines.builder.pipe.pipe_func_spec import PipeFuncSpec
 from pipelex.pipe_operators.func.pipe_func_blueprint import PipeFuncBlueprint
