+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/pipe_operator/pipe_llm/test_data.py
@@ -1,7 +1,7 @@
 from typing import ClassVar
 
 from pipelex.cogt.llm.llm_setting import LLMSetting
-from pipelex.core.pipes.pipe_input_blueprint import InputRequirementBlueprint
+from pipelex.core.pipes.input_requirement_blueprint import InputRequirementBlueprint
 from pipelex.libraries.pipelines.builder.pipe.pipe_llm_spec import PipeLLMSpec
 from pipelex.pipe_operators.llm.pipe_llm_blueprint import PipeLLMBlueprint
