+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/test_data_pipe.py
@@ -1,7 +1,7 @@
 from typing import ClassVar
 
+from pipelex.core.pipes.input_requirement_blueprint import InputRequirementBlueprint
 from pipelex.core.pipes.pipe_blueprint import PipeBlueprint
-from pipelex.core.pipes.pipe_input_blueprint import InputRequirementBlueprint
 from pipelex.libraries.pipelines.builder.pipe.pipe_signature import PipeSpec
