+++ tests/unit/pipelex/libraries/pipelines/builder/pipe/test_inputs_blueprint.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pipelex.core.pipes.pipe_input_blueprint import InputRequirementBlueprint
+from pipelex.core.pipes.input_requirement_blueprint import InputRequirementBlueprint
 
 from .test_data_inputs import InputRequirementTestCases
